requirements.txt
###### Requirements ######

binance
reqests 
python_binance
python-dotenv